Fix fields
Coordinates are now stored in DTO, so no additional data is revealed to front-end.
Description is added.
Constructors are created to be able to create empty DTOs in case of empty search results without using or modifying converter service.
If null check is implemented in converter, NoArgs and AllArgs could be deleted. However, it is not required.